Added Level-10 feature, GUI
