--- src/main/java/commands/ByeCommand.java
@@ -1,19 +1,0 @@
-package commands;
-
-import storage.Storage;
-import tasks.TaskList;
-import ui.Ui;
-
-/**
- * Bye command to exit Duke-Gerty.
- */
-public class ByeCommand extends Command {
-    @Override
-    public void execute(TaskList tl, Ui ui, Storage s) {
-        ui.display("Bye. Hope to see you again soon!");
-    }
-    @Override
-    public boolean isExit() {
-        return true;
-    }
-}
