--- src/main/java/commands/DeleteCommand.java
@@ -1,33 +1,0 @@
-package commands;
-
-import static ui.Ui.LS;
-
-import storage.Storage;
-import tasks.TaskList;
-import ui.Ui;
-
-/**
- * Delete command to handle task deleting logic.
- */
-public class DeleteCommand extends Command {
-    private int index;
-
-    /**
-     * Constructor to create DeleteCommand object.
-     */
-    public DeleteCommand(String message) {
-        index = Integer.parseInt(message) - 1;
-    }
-    @Override
-    public void execute(TaskList tl, Ui ui, Storage s) {
-        String taskText = tl.toText(this.index);
-        String taskString = tl.toString(this.index);
-        tl.removeTask(this.index);
-        s.deleteTask(taskText);
-        ui.display("Noted. I've removed this task:" + LS + taskString + LS + tl.numTasksMsg());
-    }
-    @Override
-    public boolean isExit() {
-        return false;
-    }
-}
