+++ src/main/java/duke/commands/Command.java
@@ -1,0 +1,14 @@
+package duke.commands;
+
+import duke.storage.Storage;
+import duke.tasks.TaskList;
+import duke.ui.Ui;
+
+/**
+ * Main Command class to be inherited from.
+ */
+public abstract class Command {
+    public abstract void execute(TaskList tl, Ui ui, Storage s);
+    public abstract boolean isExit();
+
+}
