+++ src/main/java/duke/commands/DeleteCommand.java
@@ -1,0 +1,33 @@
+package duke.commands;
+
+import static duke.ui.Ui.LS;
+
+import duke.storage.Storage;
+import duke.tasks.TaskList;
+import duke.ui.Ui;
+
+/**
+ * Delete command to handle task deleting logic.
+ */
+public class DeleteCommand extends Command {
+    private int index;
+
+    /**
+     * Constructor to create DeleteCommand object.
+     */
+    public DeleteCommand(String message) {
+        index = Integer.parseInt(message) - 1;
+    }
+    @Override
+    public void execute(TaskList tl, Ui ui, Storage s) {
+        String taskText = tl.toText(this.index);
+        String taskString = tl.toString(this.index);
+        tl.removeTask(this.index);
+        s.deleteTask(taskText);
+        ui.display("Noted. I've removed this task:" + LS + taskString + LS + tl.numTasksMsg());
+    }
+    @Override
+    public boolean isExit() {
+        return false;
+    }
+}
