+++ src/main/java/duke/exceptions/DukeException.java
@@ -1,0 +1,10 @@
+package duke.exceptions;
+
+/**
+ * Custom exception for duke.Duke-Gerty.
+ */
+public class DukeException extends Exception {
+    public DukeException(String errMsg) {
+        super(errMsg);
+    }
+}
