--- src/main/java/exceptions/DukeException.java
@@ -1,10 +1,0 @@
-package exceptions;
-
-/**
- * Custom exception for Duke-Gerty.
- */
-public class DukeException extends Exception {
-    public DukeException(String errMsg) {
-        super(errMsg);
-    }
-}
