+++ src/test/java/tasks/TaskListTest.java
@@ -1,9 +1,13 @@
 package tasks;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static ui.Ui.LS;
+import static duke.ui.Ui.LS;
 
 import org.junit.jupiter.api.Test;
+
+import duke.tasks.Task;
+import duke.tasks.TaskList;
+import duke.tasks.ToDo;
 
 public class TaskListTest {
     @Test
